Создание общего родительского элемента Component и наследование сущностей Tas и TasEdit от него
